requirements.txt
fastapi==0.128.0
uvicorn[standard]==0.40.0
sqlalchemy
asyncpg
python-dotenv
alembic
PyJWT
Pillow
scikit-learn
joblib
pandas
numpy
scipy
pyevmasm
xgboost
